Buncha stuff
Started making the other levels for my hook game.

Have a concept for the final project. May do a multiplayer game where arrows and WASD are used to move multiple players into rings before they hit the ground and also include some music because why not.

Need to work on a script that move the rings downward one at a time.

Also need a script for checking whether or not a player is under said ring when it hits the floor and how that interacts with the players win condition.

For hook game need a proper goal functioning.